Provide fileRequest event for code instrumentation.
Warning: This is a WIP -- this revision's API won't ship.

The final implementation will provide something like a
registry of fileRequest hooks that the client can request.

The idea is for the client to control if the code
should be processed or not during `client.createBatch()`.

For this commit:

If an event listener exists for the Hub's fileRequest event,
it will be fired for every file requested from the Hub with
the filename, file contents (as a UTF-8 Buffer) and a callback.

The listener must call the callback with a UTF-8 Buffer as the
file contents before the Hub will fufill the request.

The listener may instrument JS code using this method by
passing the provided Buffer to an instrumenter. The result
would be passed to the callback.